+++ statuspiechart.h
@@ -9,7 +9,7 @@ class StatusPieChart : public QQuickPaintedItem
 {
     Q_OBJECT
     Q_PROPERTY(QColor color READ color WRITE setColor NOTIFY colorChanged)
-    Q_PROPERTY(qreal fillAngle READ fillAngle WRITE setFillAngle NOTIFY fillAngleChanged)
+    Q_PROPERTY(int fillAngle READ fillAngle WRITE setFillAngle NOTIFY fillAngleChanged)
 
     QML_ELEMENT
 public:
@@ -18,8 +18,8 @@ public:
     QColor color() const;
     void setColor(const QColor &newColor);
 
-    qreal fillAngle() const;
-    void setFillAngle(qreal newAngle);
+    int fillAngle() const;
+    void setFillAngle(int newAngle);
 
     StatusPieChart();
 signals:
@@ -27,7 +27,7 @@ signals:
     void fillAngleChanged();
 private:
     QColor m_color;
-    qreal m_fillAngle;
+    int m_fillAngle;
 };
 
 #endif // STATUSPIECHART_H
